#CRA-459 Problem with indexes is fixed: store the trend indexes in SyntWellData instead of the trend values. -New class variables in SyntWellData with related get functions -Changes in FaciesProb and PosteriorElasticPDF related to this
git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1594 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/posteriorelasticpdf3d.h
@@ -34,7 +34,7 @@ public:
   PosteriorElasticPDF3D(const std::vector<double>            & d1, // first dimension of data points
                  const std::vector<double>                   & d2, // second dimension of data points
                  const std::vector<double>                   & d3, // third dimension of data points
-                 const std::vector<double>                   & t1, // fourth dimension (trend parameters)
+                 const std::vector<int>                      & t1, // fourth dimension (trend parameters)
                  const std::vector<std::vector<double> >     & v,  // Transformation of elastic variables from 3D to 2D
                  const double                     *const*const sigma, // Gaussian smoothing kernel in 2D
                  int                                           n1,    // resolution of density grid in elastic dimension 1
