+++ src/posteriorelasticpdf4d.h
@@ -20,8 +20,8 @@ public:
   PosteriorElasticPDF4D(const std::vector<double>                   & d1, // first dimension of data points
                         const std::vector<double>                   & d2, // second dimension of data points
                         const std::vector<double>                   & d3, // third dimension of data points
-                        const std::vector<double>                   & t1, // fourth dimension (trend parameters)
-                        const std::vector<double>                   & t2, // fourth dimension (trend parameters)
+                        const std::vector<int>                      & t1, // fourth dimension (trend parameters)
+                        const std::vector<int>                      & t2, // fourth dimension (trend parameters)
                         const std::vector<std::vector<double> >     & v,  // Transformation of elastic variables from 3D to 2D
                         const double                     *const*const sigma, // Gaussian smoothing kernel in 2D
                         int                                           n1,    // resolution of density grid in elastic dimension 1
